+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.12.0] - 2018-01-29
 
 ### Added
-- Adds menu location to the menu helper data. This helps developers to filter menu by menu `location`.
+- Menu location to the menu helper data.
+- Filters for menu object and menu items by the menu location.
 
 ## [1.11.0] - 2018-01-26